FEATURE: added a utility with a menu for selecting tasks
this commit adds a utility (`main.py`) to the root of the directory.

this utility is a menu that allows the user to select from the different tasks, and then performs the task on whatever path(s) were passed to the program.

this commit is a squashed commit of 8 former commits, titled:
- UPDATE: added a template menu utility
- UPDATE: implemented the first module to the menu utility
    - module was `data-import-export`
- FIX: files that are not CSV files no longer proceed to menu options
- UPDATE: implemented the second module to the menu utility
    - module was `data-merge-multifile`
- UPDATE: implemented the third module to the menu utility
    - module was `data-cleaner`
- UPDATE: implemented the fourth module to the menu utility
    - module was `date-to-column`
- UPDATE: implemented the fifth module to the menu utility
    - module was `split-columns-combine`
- FIX: removed `task_count` to more easily manage the menu loop